src/generators/templates/python/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pydantic[email]==2.5.0
bcrypt==4.0.1
pyjwt==2.8.0
python-multipart==0.0.6

